DeeplabV3/requirements.txt
torch~=1.13.0+cu117
torchvision~=0.14.0+cu117
numpy~=1.23.4
pillow~=9.3.0
scikit-learn~=1.2.2
tqdm~=4.32.2
matplotlib~=3.6.2
visdom~=0.2.4
PyYAML~=6.0
labelme~=5.1.1
imgviz~=1.7.2
opencv-python~=4.6.0.66